app/requirements.txt
Flask==2.0
Flask-MySQL==1.5.2  
PyMySQL==0.9.3
uWSGI==2.0.17.1
virtualenv
gunicorn
mysql-connector-python
google.cloud.secret.manager
pandas
python-dotenv
